requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
diffusers==0.28.2
torch==2.1.2
python-dotenv==1.0.1
Pillow==10.0.0